+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+- 2026-??-?? 2.3.22
+    - added Reading.protocol
+    - adding Auto-Raman to BLEDevice
+    - XS
+        - disable detector timeout during area scan
 - 2026-04-09 2.3.21
     - XS Area Scan
         - fixed line increment